api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.116.2
uvicorn[standard]==0.35.0
pydantic==2.11.9
pydantic-settings==2.7.1
pandas==2.3.2
kagglehub==0.3.13
numpy==1.26.4
torch==2.2.2+cpu

sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.2.1

bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.3.0

slowapi==0.1.9